backend/pyproject.toml
[project]
name = "pickleiq-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.135.2",
    "uvicorn",
    "pydantic>=2.0",
    "pytest-asyncio>=0.21.0",
    "psycopg[binary,pool]>=3.1",
    "structlog>=24.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
